clean up and refactor the generate_time_averages test dir t/ into the fre-python-tools tests/
beginning work on basic pytest script for generate_time_averages

cleaning up some CLI examples: approach is to generate the cdo timeaverages with the CLI
tool, then compare the time averages created with the python cdo bindings to them, and
make sure they are identical.

would be better to include some timeaveraging from a different tool e.g. fre-nctools'
timavg.csh, but that means putting it in the conda env for portability... not
desirable ? will think on this more